Renamed DTO validators
